+++ providers/src/airflow/providers/edge/cli/edge_command.py
@@ -95,6 +95,30 @@ def _pid_file_path(pid_file: str | None) -> str:
     return cli_utils.setup_locations(process=EDGE_WORKER_PROCESS_NAME, pid=pid_file)[0]
 
 
+def _write_pid_to_pidfile(pid_file_path: str):
+    """Write PIDs for Edge Workers to disk, handling existing PID files."""
+    if Path(pid_file_path).exists():
+        # Handle existing PID files on disk
+        logger.info("An existing PID file has been found: %s.", pid_file_path)
+        pid_stored_in_pid_file = read_pid_from_pidfile(pid_file_path)
+        if os.getpid() == pid_stored_in_pid_file:
+            raise SystemExit("A PID file has already been written")
+        else:
+            # PID file was written by dead or already running instance
+            if psutil.pid_exists(pid_stored_in_pid_file):
+                # case 1: another instance uses the same path for its PID file
+                raise SystemExit(
+                    f"The PID file {pid_file_path} contains the PID of another running process. "
+                    "Configuration issue: edge worker instance must use different PID file paths!"
+                )
+            else:
+                # case 2: previous instance crashed without cleaning up its PID file
+                logger.info("PID file is orphaned. Cleaning up.")
+                remove_existing_pidfile(pid_file_path)
+    logger.debug("PID file written to %s.", pid_file_path)
+    write_pid_to_pidfile(pid_file_path)
+
+
 @dataclass
 class _Job:
     """Holds all information for a task/job to be executed as bundle."""
@@ -118,7 +142,7 @@ class _EdgeWorkerCli:
 
     def __init__(
         self,
-        pid_file_path: Path,
+        pid_file_path: str,
         hostname: str,
         queues: list[str] | None,
         concurrency: int,
@@ -155,7 +179,7 @@ class _EdgeWorkerCli:
             if "404:NOT FOUND" in str(e):
                 raise SystemExit("Error: API endpoint is not ready, please set [edge] api_enabled=True.")
             raise SystemExit(str(e))
-        write_pid_to_pidfile(self.pid_file_path)
+        _write_pid_to_pidfile(self.pid_file_path)
         signal.signal(signal.SIGINT, _EdgeWorkerCli.signal_handler)
         try:
             while not _EdgeWorkerCli.drain or self.jobs: