+++ providers/tests/edge/cli/test_edge_command.py
@@ -16,6 +16,8 @@
 # under the License.
 from __future__ import annotations
 
+import logging
+import os
 from datetime import datetime
 from pathlib import Path
 from subprocess import Popen
@@ -25,10 +27,7 @@ import pytest
 import time_machine
 
 from airflow.exceptions import AirflowException
-from airflow.providers.edge.cli.edge_command import (
-    _EdgeWorkerCli,
-    _Job,
-)
+from airflow.providers.edge.cli.edge_command import _EdgeWorkerCli, _Job, _write_pid_to_pidfile
 from airflow.providers.edge.models.edge_job import EdgeJob
 from airflow.providers.edge.models.edge_worker import EdgeWorker, EdgeWorkerState
 from airflow.utils.state import TaskInstanceState
@@ -36,6 +35,44 @@ from airflow.utils.state import TaskInstanceState
 from tests_common.test_utils.config import conf_vars
 
 pytest.importorskip("pydantic", minversion="2.0.0")
+
+
+def test_write_pid_to_pidfile_success(caplog, tmp_path):
+    with caplog.at_level(logging.DEBUG):
+        pid_file_path = tmp_path / "file.pid"
+        _write_pid_to_pidfile(pid_file_path)
+        assert pid_file_path.exists()
+        assert "An existing PID file has been found" not in caplog.text
+
+
+def test_write_pid_to_pidfile_called_twice(tmp_path):
+    pid_file_path = tmp_path / "file.pid"
+    _write_pid_to_pidfile(pid_file_path)
+    with pytest.raises(SystemExit, match=r"A PID file has already been written"):
+        _write_pid_to_pidfile(pid_file_path)
+    assert pid_file_path.exists()
+
+
+def test_write_pid_to_pidfile_created_by_other_instance(tmp_path):
+    # write a PID file with the PID of this process
+    pid_file_path = tmp_path / "file.pid"
+    _write_pid_to_pidfile(pid_file_path)
+    # write a PID file, but set the current PID to 0
+    with patch("os.getpid", return_value=0):
+        with pytest.raises(SystemExit, match=r"contains the PID of another running process"):
+            _write_pid_to_pidfile(pid_file_path)
+
+
+def test_write_pid_to_pidfile_created_by_crashed_instance(tmp_path):
+    # write a PID file with process ID 0
+    with patch("os.getpid", return_value=0):
+        pid_file_path = tmp_path / "file.pid"
+        _write_pid_to_pidfile(pid_file_path)
+        assert "0" == pid_file_path.read_text().strip()
+    # write a PID file with the current process ID, call should not raise an exception
+    _write_pid_to_pidfile(pid_file_path)
+    assert str(os.getpid()) == pid_file_path.read_text().strip()
+
 
 # Ignore the following error for mocking
 # mypy: disable-error-code="attr-defined"
@@ -83,7 +120,7 @@ class TestEdgeWorkerCli:
 
     @pytest.fixture
     def worker_with_job(self, tmp_path: Path, dummy_joblist: list[_Job]) -> _EdgeWorkerCli:
-        test_worker = _EdgeWorkerCli(tmp_path / "dummy.pid", "dummy", None, 8, 5, 5)
+        test_worker = _EdgeWorkerCli(str(tmp_path / "dummy.pid"), "dummy", None, 8, 5, 5)
         test_worker.jobs = dummy_joblist
         return test_worker