+++ providers/src/airflow/providers/edge/CHANGELOG.rst
@@ -27,6 +27,22 @@
 Changelog
 ---------
 
+0.2.2re0
+.........
+
+Misc
+~~~~
+
+* ``Fixed type confusion for PID file paths (#43308)``
+
+0.2.1re0
+.........
+
+Misc
+~~~~
+
+* ``Fixed handling of PID files in Edge Worker (#43153)``
+
 0.2.0pre0
 .........